add => seperate calculators fron Index.jsx
